Improve PXLAlbum, PXLPdpAlbum construction Update README
